// modules/@h3/report/basics/enum/date-type.ts
/**
 * 日期格式类型
 */
// eslint-disable-next-line no-shadow
export enum DateFormatType {
  DEFAULT = 'default',
  Type1 = 'Type1', // 2010年5月7日
  Type2 = 'Type2', // 2010年5月
  Type3 = 'Type3', // 5月7日
  Type4 = 'Type4', // 2010/5/7
  Type5 = 'Type5', // 2010/5/7 0:00
  Type6 = 'Type6', // 10/5/7
  Type7 = 'Type7', // 5/7
  Type8 = 'Type8', // 7-May-2010
  Type9 = 'Type9', // May-2010
}